Add new equipment mode: defensive mode
The mode consists of equipping the Tank Mode equipment, but without using utamo.
This is done for cases when the user estimates they need the extra defense, but
don't want to waste resources on utamo.

An important aspect about the mode, is that when it is activated, either of
emergency or Tank mode are cancelled (utamo is cancelled) and defensive mode
is activated.

However, if HP goes into critical zone, then automatic emergency mode will
override the defensive mode.